requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile pyproject.toml
#
absl-py==2.0.0
    # via
    #   tensorboard
    #   tensorflow-cpu
alembic==1.13.0
    # via optuna
appdirs==1.4.4
    # via yfinance
astunparse==1.6.3
    # via tensorflow-cpu
beautifulsoup4==4.12.2
    # via yfinance
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via stockforecaster (pyproject.toml)
colorlog==6.8.0
    # via optuna
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flatbuffers==23.5.26
    # via tensorflow-cpu
fonttools==4.46.0
    # via matplotlib
frozendict==2.3.10
    # via yfinance
gast==0.5.4
    # via tensorflow-cpu
google-auth==2.25.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow-cpu
greenlet==3.0.2
    # via sqlalchemy
grpcio==1.60.0
    # via
    #   tensorboard
    #   tensorflow-cpu
h5py==3.10.0
    # via tensorflow-cpu
html5lib==1.1
    # via yfinance
idna==3.6
    # via requests
joblib==1.3.2
    # via scikit-learn
keras==2.14.0
    # via tensorflow-cpu
kiwisolver==1.4.5
    # via matplotlib
libclang==16.0.6
    # via tensorflow-cpu
lxml==4.9.3
    # via
    #   pytrends
    #   yfinance
mako==1.3.0
    # via alembic
markdown==3.5.1
    # via tensorboard
markupsafe==2.1.3
    # via
    #   mako
    #   werkzeug
matplotlib==3.8.2
    # via
    #   seaborn
    #   stockforecaster (pyproject.toml)
ml-dtypes==0.2.0
    # via tensorflow-cpu
multitasking==0.0.11
    # via yfinance
numpy==1.26.2
    # via
    #   contourpy
    #   h5py
    #   matplotlib
    #   ml-dtypes
    #   opt-einsum
    #   optuna
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   stockforecaster (pyproject.toml)
    #   ta
    #   tensorboard
    #   tensorflow-cpu
    #   yfinance
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow-cpu
optuna==3.5.0
    # via stockforecaster (pyproject.toml)
packaging==23.2
    # via
    #   matplotlib
    #   optuna
    #   tensorflow-cpu
pandas==2.1.4
    # via
    #   pytrends
    #   seaborn
    #   stockforecaster (pyproject.toml)
    #   ta
    #   yfinance
peewee==3.17.0
    # via yfinance
pillow==10.1.0
    # via matplotlib
protobuf==4.25.1
    # via
    #   tensorboard
    #   tensorflow-cpu
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.1.1
    # via matplotlib
pyqt5==5.15.2
    # via stockforecaster (pyproject.toml)
pyqt5-sip==12.13.0
    # via pyqt5
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytrends==4.9.2
    # via stockforecaster (pyproject.toml)
pytz==2023.3.post1
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.1
    # via optuna
requests==2.31.0
    # via
    #   pytrends
    #   requests-oauthlib
    #   tensorboard
    #   yfinance
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scikit-learn==1.3.2
    # via stockforecaster (pyproject.toml)
scipy==1.11.4
    # via scikit-learn
seaborn==0.13.0
    # via stockforecaster (pyproject.toml)
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   html5lib
    #   python-dateutil
    #   tensorboard
    #   tensorflow-cpu
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.23
    # via
    #   alembic
    #   optuna
ta==0.11.0
    # via stockforecaster (pyproject.toml)
tensorboard==2.14.1
    # via tensorflow-cpu
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow-cpu==2.14.1
    # via stockforecaster (pyproject.toml)
tensorflow-estimator==2.14.0
    # via tensorflow-cpu
tensorflow-io-gcs-filesystem==0.34.0
    # via tensorflow-cpu
termcolor==2.4.0
    # via tensorflow-cpu
threadpoolctl==3.2.0
    # via scikit-learn
tqdm==4.66.1
    # via optuna
typing-extensions==4.9.0
    # via
    #   alembic
    #   sqlalchemy
    #   tensorflow-cpu
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via requests
webencodings==0.5.1
    # via html5lib
werkzeug==3.0.1
    # via tensorboard
wheel==0.42.0
    # via astunparse
wrapt==1.14.1
    # via tensorflow-cpu
yfinance==0.2.33
    # via stockforecaster (pyproject.toml)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
